Refactor vertical coordinate funcs to new module
