+++ external/vcm/vcm/calc/constants.py
@@ -1,0 +1,20 @@
+# following are defined as in FV3GFS model (see FV3/fms/constants/constants.f90)
+_GRAVITY = 9.80665  # m /s2
+_RDGAS = 287.05  # J / K / kg
+_RVGAS = 461.5  # J / K / kg
+_LATENT_HEAT_VAPORIZATION_0_C = 2.5e6  # J/kg
+_LATENT_HEAT_FUSION = 3.3358e5  # J/kg
+_SPECIFIC_ENTHALPY_LIQUID = 4185.5
+_SPECIFIC_ENTHALPY_VAP0R = 1846
+_SPECIFIC_HEAT_CONST_PRESSURE = 1004
+_FREEZING_TEMPERATURE = 273.15
+_POISSON_CONST = 0.2854
+_EARTH_RADIUS = 6.3712e6  # m
+
+_DEFAULT_SURFACE_TEMPERATURE = _FREEZING_TEMPERATURE + 15
+
+_REFERENCE_SURFACE_PRESSURE = 100000  # reference pressure for potential temp [Pa]
+
+_SEC_PER_DAY = 86400
+_KG_M2S_TO_MM_DAY = (1e3 * 86400) / 997.0
+_KG_M2_TO_MM = 1000.0 / 997
